// src/app/core/models/Parameter.ts
export class Parameter  {
    id: number | null = null;
    folder: string | null = null;
    ranking: number | null = null;
    value: string | null = null;
    icon: string | null = null;
    color: string | null = null;
    magnitude: string | null = null;
    kind: string | null = null;
    title: string | null = null;
    description: string | null = null;
    status?: boolean | null = null;
}